refactor: rename `pkg` directory to `internal`
