+++ cmd/root.go
@@ -6,7 +6,7 @@ import (
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 
-	"tfversion/pkg/helpers"
+	"tfversion/internal/helpers"
 )
 
 var (
