+++ cmd/unalias.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 
-	"tfversion/pkg/helpers"
-	"tfversion/pkg/unalias"
+	"tfversion/internal/helpers"
+	"tfversion/internal/unalias"
 )
 
 const (
