+++ cmd/use.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"tfversion/pkg/helpers"
-	"tfversion/pkg/use"
+	"tfversion/internal/helpers"
+	"tfversion/internal/use"
 )
 
 const (
